% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{a function about knn}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{dataframe used to predict output}

\item{cl}{class value of training dataset}

\item{k_nn}{a numeric value that  specify the quantity of neighbours}

\item{k_cv}{a numeric value specify the number of folding}
}
\value{
a table with estimation, standard error, t.value and probability
that greater than t.value
}
\description{
The function predicts output class species using covariates
}
\examples{
data(my_penguins)
my_penguins <- tidyr::drop_na(my_penguins)
my_knn_cv(my_penguins[, 3:6], my_penguins$species, 1, 5)


}
\keyword{inference}
