% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{a linear model fit function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula that is used to do the linear model}

\item{data}{the data that is going to be used}
}
\value{
a table with estimation, standard error, t.value and probability
that greater than t.value
}
\description{
This fucntion fits a linear model from the two input
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)


}
\keyword{inference}
