% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{a function for cross-validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{a numeric value specify the number of folds}
}
\value{
a numeric with the cross-validation error
}
\description{
This function predict output body_mass_g using covariates bill_length_mm,
bill_depth_mm, and flipper_length_mm
}
\examples{
data(my_penguins)
my_rf_cv(5)

}
\keyword{prediction}
