% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{a t.test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a data vector with numeric values}

\item{alternative}{a string to specify the test(greater, less or two.sided)}

\item{mu}{a value of mean of null hypothesis}
}
\value{
a list, with mean of null hypothesis \code{mu}, and test type
given by \code{alternative} and a numeric p_value
}
\description{
This fucntion do a t.test with three input parameters
}
\examples{
my_t.test(c(4.2, 5.2, 6.2, 7.2, 8.2, 9.4, 10.3, 11.3, 14.2), "two.sided", 0)

}
\keyword{inference}
